Move referral view components into a separate folder with base class
