+++ app/components/referrals/personal_details_component.html.erb
