+++ app/components/referrals/previous_misconduct_component.html.erb
