+++ app/components/referrals/public_allegation_component.html.erb
