+++ app/components/referrals/public_allegation_component.rb
@@ -1,0 +1,43 @@
+module Referrals
+  class PublicAllegationComponent < ReferralFormBaseComponent
+    def rows
+      summary_rows [
+                     allegation_details_format_row,
+                     allegation_details_description_row,
+                     allegation_details_considerations_row
+                   ].compact
+    end
+
+    private
+
+    def allegation_details_format_row
+      {
+        label: "How do you want to give details about the allegation?",
+        visually_hidden_text: "how you want to give details about the allegation",
+        value: allegation_details_format(referral),
+        path: :allegation_details
+      }
+    end
+
+    def allegation_details_description_row
+      {
+        label: "Description of the allegation",
+        value: allegation_details(referral),
+        path: :allegation_details
+      }
+    end
+
+    def allegation_details_considerations_row
+      {
+        label: "Details about how this complaint has been considered",
+        value:
+          simple_format(nullable_value_to_s(referral.allegation_consideration_details.presence)),
+        path: :allegation_considerations
+      }
+    end
+
+    def section
+      Referrals::Sections::AllegationSection.new(referral:)
+    end
+  end
+end
