+++ app/components/referrals/what_happened_component.html.erb
