+++ spec/components/referrals/evidence_component_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe EvidenceComponent, type: :component do
+RSpec.describe Referrals::EvidenceComponent, type: :component do
   subject(:component) { described_class.new(referral:) }
 
   let(:row_labels) { component.rows.map { |row| row.dig(:key, :text) } }
