+++ spec/components/referrals/organisation_component_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe OrganisationComponent, type: :component do
+RSpec.describe Referrals::OrganisationComponent, type: :component do
   subject(:component) { described_class.new(referral:) }
 
   let(:row_links) { component.rows.map { |row| row.dig(:actions, :href) } }
