+++ spec/components/referrals/public_allegation_component_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 require "action_view"
 
-RSpec.describe PublicAllegationComponent, type: :component do
+RSpec.describe Referrals::PublicAllegationComponent, type: :component do
   subject(:component) { described_class.new(referral:) }
 
   let(:referrer) { referral.referrer }
